speech-streamlit/requirements.txt
altair==4.2.0
attrs==22.1.0
blinker==1.5
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
confection==0.0.3
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
entrypoints==0.4
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
importlib-metadata==5.1.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.6.2
murmurhash==1.0.9
numpy==1.23.5
packaging==22.0
pandas==1.5.2
pathy==0.10.1
Pillow==9.3.0
preshed==3.0.8
protobuf==3.20.3
pyarrow==10.0.1
pydantic==1.10.2
pydeck==0.8.0
pydub==0.25.1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
spacy==3.4.4
spacy-legacy==3.0.10
spacy-loggers==1.0.4
spacy-streamlit==1.0.4
SpeechRecognition==3.9.0
srsly==2.4.5
streamlit==1.16.0
thinc==8.1.5
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
wasabi==0.10.1
watchdog==2.2.0
wordcloud==1.8.2.2
zipp==3.11.0
ffmpeg==1.4
ffmpeg-media-converter==0.2.1
ffmpeg-python==0.2.0
ffmpeg-win64==0.0.2